+++ lib/pavlov.rb
@@ -19,15 +19,15 @@ module Pavlov
   private
 
   def self.class_for_command command_name
-    OperationFinder.find(Commands, command_name)
+    Pavlov::OperationFinder.find(Commands, command_name)
   end
 
   def self.class_for_interactor interactor_name
-    OperationFinder.find(Interactors, interactor_name)
+    Pavlov::OperationFinder.find(Interactors, interactor_name)
   end
 
   def self.class_for_query query_name
-    OperationFinder.find(Queries, query_name)
+    Pavlov::OperationFinder.find(Queries, query_name)
   end
 end
 
@@ -35,6 +35,7 @@ require_relative 'pavlov/engine' if defined?(Rails)
 require_relative 'pavlov/operation_finder'
 require_relative 'pavlov/helpers'
 require_relative 'pavlov/access_denied'
+require_relative 'pavlov/backend'
 require_relative 'pavlov/operation'
 require_relative 'pavlov/command'
 require_relative 'pavlov/query'